Aggiunge il CPI (Centro per l'Impiego) ai rapporti di lavoro

Associa a ciascun rapporto il codice e il nome del CPI a partire dal codice Belfiore della sede di lavoro (comune_sede_lavoro) o della residenza del lavoratore (comune_lavoratore). Il mapping è fornito da un lookup esplicito oppure, in sua assenza, dalle utilità geografiche di longworkR (longworkR::add_cpi_via_belfiore()), che il pacchetto non reimplementa. Le colonne sono aggiunte a dt per riferimento.

Usage

add_cpi(dt, geo = c("sede_lavoro", "residenza"), lookup = NULL, quiet = TRUE)

Arguments

dt

Un data.table di rapporti con la colonna geografica richiesta da geo. Viene modificato per riferimento.

geo

Geografia di riferimento: "sede_lavoro" (default, colonna comune_sede_lavoro) o "residenza" (colonna comune_lavoratore).

lookup

data.table o data.frame opzionale con le colonne belfiore (codice Belfiore del comune, univoco), cpi_code e, facoltativa, cpi_name (se assente viene riempita con cpi_code). NULL (default) usa longworkR, che deve essere installato e deve trovare i file di mapping in SHARED_DATA_DIR/maps. Vedi cpi_esempio per un lookup di esempio.

quiet

Se TRUE (default) l'output a console di longworkR viene soppresso; ignorato con un lookup esplicito.

Value

dt, restituito invisibilmente, con le colonne character cpi_code e cpi_name aggiunte o sovrascritte per riferimento. L'attributo ccnlcob_cpi è una lista con geo e source ("lookup" oppure "longworkR").

Details

Con lookup = NULL il mapping viene costruito una sola volta sui codici Belfiore distinti e non mancanti della colonna sorgente, non sull'intera tabella, e poi agganciato ai rapporti. Il join è un aggiornamento per riferimento: il numero e l'ordine delle righe di dt non cambiano.

Convenzioni sui residui, così che ogni rapporto resti visibile nelle distribuzioni territoriali:

codice Belfiore presente ma assente dal lookup (tipicamente comune fuori Lombardia): cpi_code = "FUORI", cpi_name = "Fuori Lombardia";

codice Belfiore mancante (NA): cpi_code = "ND", cpi_name = "Non disponibile".

Colonne cpi_code e cpi_name già presenti in dt vengono sovrascritte.

See also

Examples

library(data.table)
#> 
#> Attaching package: ‘data.table’
#> The following object is masked from ‘package:base’:
#> 
#>     %notin%
dt <- prepare_rapporti(cob_esempio)
add_cpi(dt, geo = "sede_lavoro", lookup = cpi_esempio)
dt[, .N, by = .(cpi_code, cpi_name)][order(-N)]
#>        cpi_code        cpi_name     N
#>          <char>          <char> <int>
#>  1: F205C000169      CPI MILANO  1600
#>  2: L682C000600      CPI VARESE   732
#>  3: B157C000683     CPI BRESCIA   483
#>  4: F704C000581       CPI MONZA   441
#>  5: A794C000060     CPI BERGAMO   389
#>  6: C933C000073        CPI COMO   262
#>  7:       FUORI Fuori Lombardia   250
#>  8: G388C000070       CPI PAVIA   221
#>  9: E507C000578       CPI LECCO   152
#> 10: D150C000030     CPI CREMONA   150
#> 11: E897C000034     CPI MANTOVA   132
#> 12: E648C000580        CPI LODI    99
#> 13: I829C000043     CPI SONDRIO    89
attr(dt, "ccnlcob_cpi")
#> $geo
#> [1] "sede_lavoro"
#> 
#> $source
#> [1] "lookup"
#> 

# CPI di residenza del lavoratore
add_cpi(dt, geo = "residenza", lookup = cpi_esempio)
dt[, .N, by = cpi_name][order(-N)][1:5]
#>       cpi_name     N
#>         <char> <int>
#> 1:  CPI MILANO  1429
#> 2:  CPI VARESE   746
#> 3: CPI BRESCIA   589
#> 4:   CPI MONZA   456
#> 5: CPI BERGAMO   312